Fixed test

+++ core/store/migrate/migrations/0193_s4_alter_version_type.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 
-ALTER TABLE "s4".shared ALTER COLUMN version TYPE BIGINT;
+ALTER TABLE "s4".shared ALTER COLUMN version TYPE NUMERIC;
 
 -- +goose Down
